python-service/requirements.txt
langchain==0.1.0
openai>=1.0.0
faiss-cpu==1.7.4
python-dotenv==1.0.0
nltk==3.8.1
pymongo==4.6.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
numpy>=1.24.0
scikit-learn>=1.3.0